+++ packages/algob/src/internal/deployer.ts
@@ -31,7 +31,7 @@ import type {
 	LsigInfo,
 	RuntimeEnv,
 	SCParams,
-	TxnReceipt,
+	TxReceipt,
 } from "../types";
 import { DeployerConfig } from "./deployer_cfg";
 
@@ -486,7 +486,7 @@ class DeployerBasicMode {
 	 * @param txns list transaction in group
 	 * @returns confirmed tx info of group
 	 */
-	async getReceiptTxns(txns: Transaction[]): Promise<TxnReceipt[]> {
+	async getReceiptTxns(txns: Transaction[]): Promise<TxReceipt[]> {
 		return await this.algoOp.getReceiptTxns(txns);
 	}
 }
@@ -876,7 +876,7 @@ export class DeployerDeployMode extends DeployerBasicMode implements Deployer {
 	 */
 	async executeTx(
 		transactions: wtypes.ExecParams[] | wtypes.TransactionAndSign[]
-	): Promise<TxnReceipt[]> {
+	): Promise<TxReceipt[]> {
 		return await executeTx(this, transactions);
 	}
 }
@@ -1063,7 +1063,7 @@ export class DeployerRunMode extends DeployerBasicMode implements Deployer {
 	 */
 	async executeTx(
 		transactions: wtypes.ExecParams[] | wtypes.TransactionAndSign[]
-	): Promise<TxnReceipt[]> {
+	): Promise<TxReceipt[]> {
 		return await executeTx(this, transactions);
 	}
 }